% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cut_evenly}
\alias{cut_evenly}
\title{Evenly  divide the range 0 to 1 into n pieces}
\usage{
cut_evenly(n)
}
\arguments{
\item{n}{number of points}
}
\value{
numeric vector of length n-1.
}
\description{
This function work very well in combination with \code{\link{percentify_cut}}
to create evenly spaced intervals.
}
\examples{
cut_evenly(3)
cut_evenly(4)
cut_evenly(9)

library(dplyr)

# cut_evenly() is primarily used along with percentify_cut() to space the
# cuts easily.
percentify_cut(mtcars, mpg, cut_evenly(4)) \%>\%
  summarise(mean(wt))
}
\seealso{
\link{percentify_cut}
}
