% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{even_spaced}
\alias{even_spaced}
\title{Evenly space n points between 0 and 1}
\usage{
even_spaced(n)
}
\arguments{
\item{n}{number of points}
}
\value{
numeric vector
}
\description{
Evenly space n points between 0 and 1
}
\examples{
even_spaced(3)
even_spaced(4)
even_spaced(9)
}
