% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentify.R
\name{percentify}
\alias{percentify}
\title{Title}
\usage{
percentify(data, var, lower = 0, upper = 1, key = ".percentile")
}
\arguments{
\item{data}{data.frame or tibble}

\item{var}{a variable}

\item{lower}{numeric}

\item{upper}{numeric}

\item{key}{A single character specifying the name of the virtual group
that is added. Defaults to ".percentile".}
}
\value{
percentile grouped tibble
}
\description{
Title
}
\examples{
library(dplyr)
library(broom)
percent_mtcars <- percentify_max(mtcars, mpg, c(0, 0.25, 0.5, 0.75))

percent_mtcars

summarize(percent_mtcars,
          mean_hp = mean(hp),
          mean_wt = mean(wt),
          n_obs = n()
          )

percent_mtcars \%>\%
  group_modify(~tidy(lm(disp ~ wt + cyl, data = .x)))
}
