% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentify_custom.R
\name{percentify_custom}
\alias{percentify_custom}
\title{Title}
\usage{
percentify_custom(tbl, var, lower = 0, upper = 1)
}
\arguments{
\item{tbl}{data.frame or tibble}

\item{var}{a variable}

\item{lower}{numeric}

\item{upper}{numeric}
}
\value{
percentile grouped tibble
}
\description{
Title
}
\examples{
library(dplyr)
cut_mtcars <- percentify_custom(mtcars, mpg, c(0, 0.1, 0.5), c(1, 0.6, 0.7))

summarize(cut_mtcars,
          mean_hp = mean(hp),
          mean_wt = mean(wt),
          n_obs = n()
          )
}
