% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentify_max.R
\name{percentify_max}
\alias{percentify_max}
\title{Title}
\usage{
percentify_max(tbl, var, q = numeric(), upper = 1)
}
\arguments{
\item{tbl}{data.frame or tibble}

\item{var}{a variable}

\item{q}{q}

\item{upper}{numeric}
}
\value{
percentile grouped tibble
}
\description{
Title
}
\examples{
library(dplyr)
cut_mtcars <- percentify_max(mtcars, mpg, c(0.25, 0.75))
summarize(cut_mtcars,
          mean_hp = mean(hp),
          mean_wt = mean(wt),
          n_obs = n()
          )
}
