% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentify_min.R
\name{percentify_min}
\alias{percentify_min}
\title{Title}
\usage{
percentify_min(tbl, var, q = numeric(), lower = 0)
}
\arguments{
\item{tbl}{data.frame or tibble}

\item{var}{a variable}

\item{q}{q}

\item{lower}{numeric}
}
\value{
percentile grouped tibble
}
\description{
Title
}
\examples{
library(dplyr)
library(broom)
percent_mtcars <- percentify_min(mtcars, mpg, c(0.25, 0.5, 0.75, 1))

percent_mtcars

summarize(percent_mtcars,
          mean_hp = mean(hp),
          mean_wt = mean(wt),
          n_obs = n()
          )

percent_mtcars \%>\%
  group_modify(~tidy(lm(disp ~ wt + cyl, data = .x)))
}
