% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentify_random.R
\name{percentify_random}
\alias{percentify_random}
\title{Group a data.frame by percentile ranges in variable with constant upper bound}
\usage{
percentify_random(data, var, width, n = 10)
}
\arguments{
\item{data}{A data.frame or tibble,}

\item{var}{Variable to do grouping by as string or symbol.}

\item{width}{Numerical values for the width of the interval. Must be between
0 and 1.}

\item{n}{Integer, number of intervals to create.}
}
\value{
percentile grouped tibble
}
\description{
This function is a wrapper around \code{\link{percentify}}. Given a \code{width`` and a }n`, it will generate a number of random interval with fixed width.
}
\examples{
library(dplyr)
library(broom)
percent_mtcars <- percentify_random(mtcars, mpg, 0.4, 10)

percent_mtcars

summarize(percent_mtcars,
          mean_hp = mean(hp),
          mean_wt = mean(wt),
          n_obs = n()
          )

percent_mtcars \%>\%
  group_modify(~tidy(lm(disp ~ wt + cyl, data = .x)))
}
\seealso{
Other percentile samplers: \code{\link{percentify_cut}},
  \code{\link{percentify_max}},
  \code{\link{percentify_min}}, \code{\link{percentify}}
}
\concept{percentile samplers}
