% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect.percentiled_df}
\alias{collect.percentiled_df}
\title{Force virtual groups to become explicit rows}
\usage{
collect.percentiled_df(x, ..., id = NULL, original_id = NULL)
}
\arguments{
\item{x}{A \code{percentiled_df}.}

\item{...}{Not used.}

\item{id}{Optional. A single character that specifies a name for a column
containing a sequence from \code{1:n} for each group.}

\item{original_id}{Optional. A single character that specifies a name for
a column containing the original position of the row.}
}
\description{
When \code{collect()} is used on a \code{percentiled_df}, the virtual groups
are made explicit.
}
\examples{
library(dplyr)

# virtual groups become real rows
collect(percentify_cut(mtcars, mpg, c(0.5)))
}
