% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cut_evenly}
\alias{cut_evenly}
\title{Evenly devide the range 0 to 1 into n pieces}
\usage{
cut_evenly(n)
}
\arguments{
\item{n}{number of points}
}
\value{
numeric vector
}
\description{
Evenly devide the range 0 to 1 into n pieces
}
\examples{
cut_evenly(3)
cut_evenly(4)
cut_evenly(9)

library(dplyr)

# cut_evenly() is primarily used along with percentify_cut() to space the
# cuts easily.
percentify_cut(mtcars, mpg, cut_evenly(4)) \%>\%
  summarise(mean(wt))
}
