% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentify.R
\name{percentify}
\alias{percentify}
\title{Title}
\usage{
percentify(data, var, lower = 0, upper = 1, key = ".percentile")
}
\arguments{
\item{data}{data.frame or tibble}

\item{var}{a variable}

\item{lower}{numeric}

\item{upper}{numeric}

\item{key}{A single character specifying the name of the virtual group
that is added. Defaults to ".percentile".}
}
\value{
percentile grouped tibble
}
\description{
Title
}
\examples{
library(dplyr)
cut_mtcars <- percentify(mtcars, mpg, c(0, 0.1, 0.5), c(1, 0.6, 0.7))

summarize(cut_mtcars,
          mean_hp = mean(hp),
          mean_wt = mean(wt),
          n_obs = n()
          )
}
