% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentile_cut.R
\name{percentify_cut}
\alias{percentify_cut}
\title{Group a data.frame by percentile ranges in variable by cutting points}
\usage{
percentify_cut(data, var, q = numeric())
}
\arguments{
\item{data}{A data.frame or tibble,}

\item{var}{Variable to do grouping by as string or symbol.}

\item{q}{Numerical values for cutting points. Must be between 0
and 1.}
}
\value{
percentile grouped tibble
}
\description{
This function is a small wrapper around \code{\link{percentify}}. This
function takes a vector of points and creates ranges between those points,
resulting in non-overlapping groups.
}
\details{
There is a \code{\link[ggplot2:autoplot]{ggplot2::autoplot()}} to vizualise the the percentile ranges.
}
\examples{
library(dplyr)
library(broom)
percent_mtcars <- percentify_cut(mtcars, mpg, c(0.25, 0.5, 0.75))

percent_mtcars

summarize(percent_mtcars,
          mean_hp = mean(hp),
          mean_wt = mean(wt),
          n_obs = n()
          )

percent_mtcars \%>\%
  group_modify(~tidy(lm(disp ~ wt + cyl, data = .x)))

# cut_evenly() can be used to create cuts
percentify_cut(mtcars, mpg, cut_evenly(8))

library(ggplot2)
autoplot(percent_mtcars)
}
\seealso{
\link{cut_evenly}

Other percentile samplers: \code{\link{autoplot.percentiled_df}},
  \code{\link{percentify_max}},
  \code{\link{percentify_min}},
  \code{\link{percentify_random}}
}
\concept{percentile samplers}
