% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentile_cut.R
\name{percentify_cut}
\alias{percentify_cut}
\title{Title}
\usage{
percentify_cut(tbl, var, q = numeric())
}
\arguments{
\item{tbl}{data.frame or tibble}

\item{var}{a variable}

\item{q}{q}
}
\value{
percentile grouped tibble
}
\description{
Title
}
\examples{
library(dplyr)
library(broom)
percent_mtcars <- percentify_cut(mtcars, mpg, c(0.25, 0.5, 0.75))

percent_mtcars

summarize(percent_mtcars,
          mean_hp = mean(hp),
          mean_wt = mean(wt),
          n_obs = n()
          )

percent_mtcars \%>\%
  group_modify(~tidy(lm(disp ~ wt + cyl, data = .x)))

# cut_evenly() can be used to create cuts
percentify_cut(mtcars, mpg, cut_evenly(8))
}
