% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentify_max.R
\name{percentify_max}
\alias{percentify_max}
\title{Group a data.frame by percentile ranges in variable with constant upper bound}
\usage{
percentify_max(data, var, q = numeric(), upper = 1)
}
\arguments{
\item{data}{A data.frame or tibble,}

\item{var}{Variable to do grouping by as string or symbol.}

\item{q}{Numerical values for lower bound of ranges. Must be between 0
and 1.}

\item{upper}{Numerical values for upper bound of ranges. Must be between 0
and 1. Length of lower and upper must be equal.}
}
\value{
percentile grouped tibble
}
\description{
This function is a small wrapper around \code{\link{percentify}}, where the
upper bound is a fixed value.
}
\details{
There is a \code{\link[ggplot2:autoplot]{ggplot2::autoplot()}} to visualize the the percentile ranges.
}
\examples{
library(dplyr)
library(broom)
percent_mtcars <- percentify_max(mtcars, mpg, c(0, 0.25, 0.5, 0.75))

percent_mtcars

summarize(percent_mtcars,
          mean_hp = mean(hp),
          mean_wt = mean(wt),
          n_obs = n()
          )

percent_mtcars \%>\%
  group_modify(~tidy(lm(disp ~ wt + cyl, data = .x)))

library(ggplot2)
autoplot(percent_mtcars)
}
\seealso{
Other percentile samplers: \code{\link{autoplot.percentiled_df}},
  \code{\link{percentify_cut}},
  \code{\link{percentify_min}},
  \code{\link{percentify_random}}
}
\concept{percentile samplers}
