% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentify_min.R
\name{percentify_min}
\alias{percentify_min}
\title{Title}
\usage{
percentify_min(tbl, var, q = numeric(), lower = 0)
}
\arguments{
\item{tbl}{data.frame or tibble}

\item{var}{a variable}

\item{q}{q}

\item{lower}{numeric}
}
\value{
percentile grouped tibble
}
\description{
Title
}
\examples{
library(dplyr)
cut_mtcars <- percentify_min(mtcars, mpg, c(0.25, 0.75))
summarize(cut_mtcars,
          mean_hp = mean(hp),
          mean_wt = mean(wt),
          n_obs = n()
          )
}
